+++ Flex/Message.hs
@@ -7,11 +7,12 @@
 
 module Flex.Message (LocatedMsg, Msg(..)) where
 
+import Data.Text.Lazy (Text)
 import Flex.Position
 
 
-type LocatedMsg p = (String, p)
+type LocatedMsg p = (Text, p)
 
 class (Pos p, Monad m) => Msg p m where
-  errorLexer :: p -> String -> m a
+  errorLexer :: p -> Text -> m a
   -- ^ Handler for lexing errors