+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.1.1
+
+* Replace `String` by `Text`.
+
+* Remove labels.
+
+* Rename functions in `Flex.Position`.
+
+
 ## v0.1.0
 
 Change lexing style to TeX-like line-by-line lexing.